# swagger.yaml
swagger: '2.0'
info:
  description: Mobilaw Backend
  version: 1.0.0
  title: Mobilaw API Doc
  # put the contact info for your development or API team
paths:
...
  /panel/all:
    get:
      tags:
      - admins
      summary: gets all questions asked by users
      #operationId: register
      description: gets all questions asked by users
      produces:
      - application/json 
      parameters: 
      - in: header
        name: Authorization
        description: token provided by login endpoint
        name: Authorization
        type: string
        format: jwt
      responses:
        200:
          description: search results
          schema:
            type: array
            items:
              $ref: '#/definitions/Question'
          
        403:
          description: jwt token is not valid
  /panel/all/active:
    get:
      tags:
      - admins
      summary: gets all active questions asked by users
      #operationId: register
      description: gets all active questions asked by users
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: token provided by login endpoint
        name: Authorization
        type: string
        format: jwt
      responses:
        200:
          description: search result matching criteria

          schema:
              $ref: '#/hiddenDef/QuestionActive'
        403:
          description: jwt token is not valid
  /panel/all/unactive:
    get:
      tags:
      - admins
      summary: gets all unactive questions asked by users
      #operationId: register
      description: gets all unactive questions asked by users
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: token provided by login endpoint
        name: Authorization
        type: string
        format: jwt
      responses:
        200:
          description: search result matching criteria

          schema:
              $ref: '#/hiddenDef/QuestionUnactive'
        403:
          description: jwt token is not valid
  /panel/reply/:questionId:
    post:
      tags:
      - admins
      summary: gets all questions asked by users
      #operationId: register
      description: gets all questions asked by users
      produces:
      - application/json
      parameters:
      - in: path
        name: QuestionID
        description: questionid to reply

      - in: body
        name: Reply
        description: text to reply
        schema:
          $ref: '#/hiddenDef/Reply'
      - in: header
        name: Authorization
        description: token provided by login endpoint
        name: Authorization
        type: string
        format: jwt
      responses:
        200:
          description: replied
        403:
          description: jwt token is not valid
...
definitions:
...
  Question:
    properties:
      email:
        type: string
        example: mobilaw@gmail.com
      text:
        type: string
        example: Something
      typeName:
        type: string
        example: Hukuki Soru
      user:
        type: string
        example: ObjectId("4d3ed089fb60ab534684b7ff")
      category:
        type: string
        example: ObjectId("4d3ed089fb60ab534684b7ff")
      isActive:
        type: boolean
        example: false
      answer:
        type: string
        example: Done

# Added by API Auto Mocking Plugin
schemes:
 - http